refactor confirmation handler
rename it
move it to its own package
